[restructure] remove modular widgets & themes
everythings changing now, hopefully for the better
