--- cogs/profile/__init__.py
@@ -1,7 +1,0 @@
-from . import profile
-
-def setup(bot):
-    profile.setup(bot)
-
-def teardown(bot):
-    profile.teardown(bot)
