--- cogs/profile/widgets/__init__.py
@@ -1,4 +1,0 @@
-from .base import *
-from .badges import *
-from .account_age import *
-from .join_date import *
